+++ sdk/ai/Azure.AI.Projects/src/Custom/RedTeams/RedTeams.cs
@@ -2,10 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 
 namespace Azure.AI.Projects;
 
+[Experimental("AAIP001")]
 public partial class RedTeams
 {
     /// <summary>
@@ -47,7 +49,7 @@ public partial class RedTeams
     {
         Argument.AssertNotNull(redTeam, nameof(redTeam));
 
-        using PipelineMessage message = CreateCreateRequest(redTeam, options);
+        using PipelineMessage message = CreateCreateRequest(redTeam,options);
         ClientResult result = ClientResult.FromResponse(await Pipeline.ProcessMessageAsync(message, options).ConfigureAwait(false));
         return ClientResult.FromValue((RedTeam)result, result.GetRawResponse());
     }